Separate s3, azure, updater and common
