+++ neuro_san/service/watcher/temp_networks/s3/aws_async_client_worker.py
@@ -40,7 +40,7 @@ from botocore.exceptions import NoCredentialsError
 
 from leaf_common.logging.sensitive_logger import SensitiveLogger
 
-from neuro_san.service.watcher.temp_networks.s3_util import S3Util
+from neuro_san.service.watcher.temp_networks.s3.s3_util import S3Util
 
 
 # pylint: disable=too-many-instance-attributes
