+++ neuro_san/service/watcher/temp_networks/s3/s3_reservations_expiration.py
@@ -32,9 +32,9 @@ from botocore.client import BaseClient
 from botocore.exceptions import ClientError
 
 from neuro_san.interfaces.reservationist import Reservationist
-from neuro_san.service.watcher.temp_networks.aws_sync_client_worker import AwsSyncClientWorker
-from neuro_san.service.watcher.temp_networks.s3_reservations_retriever import S3ReservationsRetriever
-from neuro_san.service.watcher.temp_networks.s3_util import S3Util
+from neuro_san.service.watcher.temp_networks.s3.aws_sync_client_worker import AwsSyncClientWorker
+from neuro_san.service.watcher.temp_networks.s3.s3_reservations_retriever import S3ReservationsRetriever
+from neuro_san.service.watcher.temp_networks.s3.s3_util import S3Util
 
 
 class S3ReservationsExpiration:
