+++ neuro_san/service/watcher/temp_networks/s3/s3_util.py
